requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
aiofiles==23.2.1
Jinja2==3.1.4
pydantic[email]==2.7.1
gunicorn==21.2.0
pytest==9.0.2
pytest-watch==4.2.0